Import nemiver-0.9.4.
Nemiver is a standalone graphical debugger that integrates well in the
GNOME desktop environment. It currently features a backend which uses
the well known GNU Debugger gdb to debug C / C++ programs.

Note that it is *not* yet working properly so it's not hooked up.

ok jasper@

Written by: Antoine Jacoutot <ajacoutot@openbsd.org>