Pull in r175057 from upstream LLVM, via FreeBSD, found by Wada Hiroyuki.
X86: Disable generation of rep;movsl when %esi is used as a base pointer.
This happens when there is both stack realignment and a dynamic alloca in the
function. If we overwrite %esi (rep;movsl uses fixed registers) we'll lose the
base pointer and the next register spill will write into oblivion.

Written by: Stuart Henderson <sthen@openbsd.org>