Import security/easy-rsa, which is split off from OpenVPN in 2.3+. From Bjorn Ketelaars, ok benoit@
Easy-RSA is a small RSA key management package, based on the openssl
command line tool, that can be found in the easy-rsa subdirectory
of the OpenVPN distribution. While this tool is primarily concerned
with key management for the SSL VPN application space, it can also
be used for building web certificates.

Written by: Stuart Henderson <sthen@openbsd.org>