To be able to use c11 standard atomics on arm, we need to tell llvm whether or not we have the atomic operations ldrex/strex.
By defining the the atomic inline width to 64, we allow llvm to
use 64-bit wide ldrex/strex instructions on arm. Those only work
on newer arm cores.

In llvm-3.3 a new check was implemented, which does check which
arm core is supported more precisely. It's only enabled for Linux,
so change it to count us in, too.