KTron is a simple snake duel game for the KDE. You can play KTron against the computer or a friend. The aim of the game is to live longer than your opponent. To do that, avoid running into a wall, your own tail and that of your opponent.
KSnakeDuel is a simple Snake-like game for the KDE. The aim of
KSnake is to survive as long as possible and eat as many fruits as
you can.

You can switch between these games using the game type selector in
the configuration dialog.

Written by: zhuk <zhuk@cvs.openbsd.org>