SMOKE is a introspective wrapper around the Qt and KDE frameworks. Legend has it that SMOKE stands for Scripting Meta Object Kompiler Engine.
Information about all classes and all methods/functions is stored
in cross-referencing tables for fast look-ups. Thereby the whole
API of a wrapped library can be used.

The main purpose of SMOKE is making it easier to write bindings
from scripting languages to Qt and KDE - with an emphasis on ease
of use and flexibility.

This package contains wrappers for Qt, and does not depend on KDE
in any way.

NB: This could conflict with x11/ruby-qt4, paths to fix this are
investigated.

Written by: zhuk <zhuk@cvs.openbsd.org>