Update to firefox 21.0.
- See http://www.mozilla.org/en-US/firefox/21.0/releasenotes/
- Fixes MFSA 2013-41->48
- switch back again to use gcc 4.6 on i386, ld hits the memory exhausted
  wall again with objects built with clang. doh.
- stuff shuffled around in PLIST

*BIG SCARY WARNING FOR PRIVACY-CONCERNED USERS*
Firefox 21 starts implementing something called 'Firefox Health Report',
which sends _anonymized_ data to mozilla servers about Firefox
performance.
This is opt-out - users will be prompted at first start if they want to
keep it enabled. This is not turned as an opt-in knob in the port on
purpose, because i got no feedback on this, and that will allow upstream
to know they have some OpenBSD users. This is of course subject to change.

See https://blog.mozilla.org/futurereleases/2013/05/14/firefox-heal-thyself/
& about:healthreport for more info about this.

Written by: Landry Breuil <landry@openbsd.org>

Conflicts:
	www/mozilla-firefox/patches/patch-js_src_configure_in
	www/mozilla-firefox/patches/patch-toolkit_library_Makefile_in