Import hs-OpenGLRaw-1.3.0.0:
OpenGLRaw is a raw Haskell binding for the OpenGL 3.2 graphics
system and lots of OpenGL extensions. It is basically a 1:1 mapping
of OpenGL's C API, intended as a basis for a nicer interface.
OpenGLRaw offers access to all necessary functions, tokens and types
plus a general facility for loading extension entries. The module
hierarchy closely mirrors the naming structure of the OpenGL
extensions, making it easy to find the right module to import.  All
API entries are loaded dynamically, so no special C header files
are needed for building this package. If an API entry is not found
at runtime, a userError is thrown.

ok sthen@

(will be hooked to the build during the ghc update)


Written by: Matthias Kilian <kili@openbsd.org>