create a special 'in-between' task that runs (setup only actually) between checksum and extract: that's when we're truely committed to building the port there.
create the affinity markers at that point, and not sooner.
since we know whether we're building in memory, we can record that.

if we're building in-memory, always ask for size, since it's very very
fast.

special treatment for tagged ports on error: mark them as "cleaned" so
the host gets untainted eventually.

Written by: Marc Espie <espie@openbsd.org>