setup of serialized tasks *is* deterministic.
The moment when we're sure we're locked in uninstall
is the right moment to revisit tainted status...
not later, because if we do that in finalize, some tainted ports
may have finished building, and we may think we're untainted now.
as seen by landry@

Note also that setup runs in the main dpb process, so not concurrently with
the engine, and so the engine can safely check for tags right after setup,
even if pkg_delete has not finished, because by the time the port can lock
and reach depends, pkg_delete must have finished.

Written by: Marc Espie <espie@openbsd.org>