fix restart issues...
first tag we see will trigger a junk, unless we already
junked on that host during this run.

that way, we can forget whether we were building kde3 or kde4,
as we start with a "clean slate".

I could have junked at start, but this is much better, as it junks
"just in time" for the tag.

Written by: Marc Espie <espie@openbsd.org>