imports textproc/py-cssselect, ok jasper@
cssselect parses CSS3 Selectors and translate them to XPath 1.0
expressions. Such expressions can be used in lxml or another XPath
engine to find the matching elements in an XML or HTML document.

This module used to live inside of lxml as lxml.cssselect before it was
extracted as a stand-alone project.

Written by: Stuart Henderson <sthen@openbsd.org>