import ports/audio/checkmate, ok weerd@
Checkmate MP3 Checker is a free program that checks MP3 files for
errors. It scans MP3 files to see if the frames are where they are
supposed to be, whether the frame headers are correct and whether the
frame headers are consistent throughout the file. It reports some
information on each file and an indication whether the file is good or
bad.

Written by: Stuart Henderson <sthen@openbsd.org>