Resurrect xfce4-datetime, version 0.6.2.
This plugin shows the date and time in the panel, and a calendar
appears when you left-click on it. It is a simpler alternative to Orage.

Somehow i resurrected it upstream 10 months ago but totally forgot to
readd it in the portstree - been using it without issues since then.

Written by: Landry Breuil <landry@openbsd.org>