Repair time/delay/duration time_t issues in gnat. All direct imports of C functions have been modified to use same-size types on both sides. There are no checks in gcc to make sure the types match, and it happily trashes the stack or reads bogus data without this fix.
The C shim in gnatlib is still casting down to long (by design), thus
gnat is not 2038-proof on 32 bit architectures. Fixing this is an
ongoing project, but I'm running out of time for this release.

Getting the timezone offset is now supported; the unused usleep
procedure that snuck in from FreeBSD support got removed.

Tested on all supported architectures, plus new bootstraps due to the
stack trashing, just in case...

Conceptually fine with brad@
deraadt@ is unhappy with the partial solution, fixing this remains a
high priority todo

Written by: tobiasu <tobiasu@cvs.openbsd.org>