Bump after recent gnome.port.mk changes; these do not require to RUN_DEPENDS on yelp.
Note that this change will drop a big pile of dependencies from the pkg
and it is possible that some dependencies will be missing from the
dependency chain; if that is the case, it means that the pkg dependencies
were wrong in the first place and would have required fixing anyway.

Any fallout, please let me know.

Written by: Antoine Jacoutot <ajacoutot@openbsd.org>