(re-)import zathura/core 0.2.9.
zathura is a highly customizable and functional PDF viewer based
on the poppler rendering library and the gtk+ toolkit. The idea
behind zathura is an application that provides a minimalistic and
space saving interface as well as an easy usage that mainly focuses
on keyboard interaction.

Note that after 0.1.0, zathura was rearchitectured/splitted between
'core' (the viewer itself) and 'plugins' (taking care of the rendering
itself). Users of previous version will need to install
zathura-pdf-poppler or zathura-pdf-mupdf to restore their featureset.

ok ajacoutot@

Written by: Landry Breuil <landry@openbsd.org>