Fix tests to properly recognize OpenBSD.
openbsdrcctl: extend enabled() to check flags
On OpenBSD, it's possible to define particular flags a daemon should be
enabled with. To make sure the flags are properly updated when changed
(in pillars or states), we return a service as "disabled" if the
currently configured flags do not match the provided ones; this ensures
/etc/rc.conf.local is always in sync.

All patches merged upstream.

Written by: Antoine Jacoutot <ajacoutot@openbsd.org>