Import multimedia/sfml, requested by and ok bcallah@
SFML is a portable and easy to use multimedia API written in C++.
You can see it as a modern, object-oriented alternative to SDL.
SFML is composed of several packages to perfectly suit your needs.
You can use SFML as a minimal windowing system to interface with
OpenGL, or as a fully-featured multimedia library for building games
or interactive programs.

Written by: Pascal Stumpf <pascal@openbsd.org>