update conserver to 8.2.0, and make some port changes ...
- Switch to using unix domain sockets instead of network sockets for
comms between "console" client and "conserver" server. This is a compile-
time toggle and disables network sockets. Uses SO_PEERCRED for access
control so it's possible to do per-user acls in conserver without
the need to re-enter user passwords.

- Add a flavour to use network sockets for those that need it (i.e.
running the client on a different machine to the server).

- Tweaks to pkg-readme explaining the above.

Written by: Stuart Henderson <sthen@openbsd.org>