Import editors/ne ok juanfra@
ne is a free (GPL'd) text editor based on the POSIX standard that runs
on almost any UN*X machine. ne is easy to use for the beginner, but
powerful and fully configurable for the wizard, and most sparing in its
resource usage.

Features:
* multiple user interfaces
* syntax highlighting
* full UTF-8 support
* simple scripting language
* binary file editor

Written by: bcallah <bcallah@cvs.openbsd.org>