Import graphics/slop ok landry@
slop (Select Operation) is an application that querys for a selection
from the user and prints the region to stdout. It grabs the mouse and
turns it into a crosshair, lets the user click and drag to make a
selection (or click on a window) while drawing a pretty box around it,
then finally prints the selection's dimensions to stdout.

Features:
* Hovering over a window will cause a selection rectangle to appear over
  it.
* Clicking on a window makes slop return the dimensions of the window.
* Clicking and dragging causes a selection rectangle to appear, renders
  pretty well (compared to scrot). And will return the dimensions of
  that rectangle in absolute screen coords.
* On startup it turns your cursor into a crosshair, then adjusts the
  cursor into angles as you drag the selection rectangle.
* Supports simple arguments:
        * Change selection rectangle border size.
        * Select X display.
        * Set padding size, even negative padding sizes!
        * Set click tolerance for if you have a shaky mouse.
        * Set the color of the selection rectangles to match your theme!
          (Even supports transparency!)
        * Remove window decorations from selections.

Written by: bcallah <bcallah@cvs.openbsd.org>