Update umfpack to 5.6.2.
Stop using AMD and hacking the UFConfig dependency (which caused all the
freemat mess).

There is no way to ditch the AMD alloctaion symbols in the last
umfpack stand-alone release so I patched them.

This is fixed in SuiteSparse which is supposed to put all the sparse
matrix libraries in one spot, but this is a large project left
for better times.

This is part 3/3 that allows freemat to build and package.

Regression tests seem to be fine on both octave and py-scipy.

First found by and discussed with sthen@

Written by: Paul Irofti <pirofti@openbsd.org>