Update to Asterisk 11.14.2, fixing AST-2014-019 in res_http_websocket, a crash when receiving a WebSocket packet with a payload length of zero.
"Frames with a payload length of 0 were incorrectly handled in
res_http_websocket. Provided a frame with a payload had been received
prior it was possible for a double free to occur. The realloc operation
would succeed (thus freeing the payload) but be treated as an error.
When the session was then torn down the payload would be freed again
causing a crash. The read function now takes this into account."

Written by: Stuart Henderson <sthen@openbsd.org>