Import devel/shunit2 OK aja@
shUnit2 is a xUnit unit test framework for Bourne based shell scripts,
and it is designed to work in a similar manner to JUnit, PyUnit, etc.
If you have ever had the desire to write a unit test for a shell script,
shUnit2 can do the job.

Written by: rpe <rpe@cvs.openbsd.org>