Workaround for what seems to be a clang issue.
This package compiles and seems to work using gcc 4.8.

sysdep.h includes ansidecl.h and for some reason the original file that
includes both doesn't see the stuff defined ansidecl.h.  By removing the
reentrancy guard it works.  Tried recreating this in a small scale
example but that worked as expected.