binutils: fix compile after gdb update
gdb provides headers in /usr/local/include.  This ports
includes it to have access to libintl.h.  At the same
time, it looks for other headers which are also present
there, even though they might be of older version.

To fix this, disable NLS support, so that we don't need
libintl.h anymore, so that we do not have to include
/usr/local/include.

"sure" marco@