import p5-Exporter-Tiny 0.042 OK benoit@ sthen@
exporter with the features of Sub::Exporter

Exporter::Tiny supports many of Sub::Exporter's external-facing
features including renaming imported functions with the -as, -prefix
and -suffix options; explicit destinations with the into option;
and alternative installers with the installler option.  But it's
written in only about 40% as many lines of code and with zero
non-core dependencies.

Written by: Alexander Bluhm <bluhm@openbsd.org>