texlive_base: Remove bogus symlink.
Although latexmk is included in texlive, we omit it in favour of our
separate port. texlive_base should not install a symlink.

There was also confusion between latex-mk and latexmk. We have both in
ports, but the latter is the one related to texlive. Let's mark is as a soft
dependency.

Upgrade path is currently leaving a partial package behind due to the bogus
symlink. A quirk will probably be required, which we can commit later.

OK ajacoutot@

Written by: Edd Barrett <edd@openbsd.org>