Import gdmap-0.8.1
GdMap is a tool which allows to visualize disk space. To display directory
structures cushion treemaps are used which visualize a complete folder or even
the whole hard drive with one picture.

Cushion treemaps display directories and files in rectangular areas. The larger
a file is the larger is the rectangle which represents it. All files in one
directory are painted within the rectangle of that directory.

OK kirby@

Written by: czarkoff <czarkoff@cvs.openbsd.org>