tweak WANTLIB again; build problem reported by ajacoutot.
quadmath is used by -f95 (WANTLIB), however the lib is also present in GCC 4.9
which is required all over the place for bulk builds. So bodge around it by
removing the WANTLIB and at least specifying a RUN_DEPENDS-f95 instead.

This is pretty much academic anyway as you can't install gcc-4.8 on a typical
workstation as it will conflict with gcc-4.9 libs that are a common dependency.

Written by: Stuart Henderson <sthen@openbsd.org>