Import direvent-5.0.
Direvent is a directory content watcher daemon, i.e. a program that monitors a
set of directories on the file system and reacts when their content changes.
When a change is detected, the daemon reacts by invoking an external command
configured for that kind of change.
<...>

ok jasper@

Written by: Antoine Jacoutot <ajacoutot@openbsd.org>