update to opam-1.2.2
Because the Aspcud constraint solver is not ported to OpenBSD yet, this
package is configured to use the IRILL online solver by default.  To
disable this and use the internal solver, set the "OPAMNOASPCUD" env
variable to "1".  You will need to do this if you want to use OPAM
when offline.

ok jca@, tested by krw@ daniel@ jsg@

Written by: Anil Madhavapeddy <avsm@openbsd.org>