import ports/mail/rspamd, joint work with brad - not yet perfect, but in good enough shape to continue in tree.  tweak/ok giovanni@.
Rspamd is an advanced spam filtering system that allows evaluation of messages
by a number of rules including regular expressions, statistical analysis and
custom services such as URL black lists. Each message is analysed by rspamd and
given a spam score.

According to this spam score and the user's settings rspamd recommends an action
for the MTA to apply to the message- for example to pass, reject or add a
header. Rspamd is designed to process hundreds of messages per second
simultaneously and has a number of features available.

Written by: Stuart Henderson <sthen@openbsd.org>