the external editline library is not required since that openbsd editline in base is OK and happy.
sha256/binaries in the rc package are equal `with and without' this change,
that means that the editline provided by external editline-1.5.tar.gz was never used,
so we can kill it with fire.

regen distfile; no package change...so no need bump.

noticed and diff from Michael Reed(MAINTAINER)