Re-introduce the suhosin extension and the no_suhosin flavor as well which can be used to disable the suhosin extension in the php packages.
We are not going to build this flavor by default because people should
use suhosin by default, but in some corner-cases if someone wants to disable
it, the flavor can be used.

Also move some variables and the sybase_ct subpackage from the Makefiles
to the shared one, there is no need to have the same data in 3 different
Makefiles.

Written by: Robert Nagy <robert@openbsd.org>