import ports/security/py-acme, ok landry@ note, this is still devel code, but useful to have in tree
ACME (Automated Certificate Management Environment) is a protocol for
automating the management of domain-validation certificates, based on a
simple JSON-over-HTTPS interface.

The spec (https://github.com/letsencrypt/acme-spec) is a work in progress
but it is hoped to eventually move it to the IETF process to become an RFC.

This packages provides a Python implementation of the protocol.

Written by: Stuart Henderson <sthen@openbsd.org>