Import brumbrumrally-0.5.
Brum Brum Rally is a simple 2D racing game with autogenerated tracks so
that you don't have to play the same tracks over and over again.

It's possible to play single races or tournaments with up to eight
players at the same time. You can play with your friends on the same
computer and/or with other people over a network (Internet or LAN).

It's also possible play alone against the computer, or against the clock
by trying to finish a race as fast as possible in the time trial mode.

ok landry@