Import py-bsddb3-6.0.1.
This module provides a nearly complete wrapping of the Oracle/Sleepycat C API
for the Database Environment, Database, Cursor, Log Cursor, Sequence and
Transaction objects, and each of these is exposed as a Python type in the
bsddb3.db module.
<...>

Note that this is not the most recent release because our databases/db/ ports
are too ancient.
ok jasper@

Written by: Antoine Jacoutot <ajacoutot@openbsd.org>