Import lttoolbox-3.3.1.
lttoolbox is a toolbox for lexical processing, morphological analysis
and generation of words. Analysis is the process of splitting a word
(e.g., cats) into its lemma 'cat' and the grammatical information
<n><pl>. Generation is the opposite process.

The package is split into three programs: lt-comp, the compiler;
lt-proc, the processor; and lt-expand, which generates all possible
mappings between surface forms and lexical forms in the dictionary.

ok jasper@